requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-Cors==3.0.8
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
pymongo==3.10.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
